fix: CSV function correctly handling all intents, var config updated
Previously the xpath function was doing a match for title text using "contains". As this is a substring search, this could cause an overwrite to occur for similar strings. For example, "None" would overwrite "No" as this is a substring match. This has now been implemented using "equals", which looks for an exact match and the issue has been resolved.

variable_names was also updated to contain all the correct intent and entity names for this project.