feat: Confusion matrix data added
Scraper can now navigate the TP, FP, TN, FN scores for the intent the batch test is based on and add these values to the scores dict. Modifications will be needed to make this work for batch tests that are now done per intent (eg. one batch test for all intents...)